+++ attrbench/functional/__init__.py
@@ -3,8 +3,6 @@ from .impact_score import impact_score
 from .infidelity import infidelity
 from .insertion_deletion import insertion, deletion
 from .max_sensitivity import max_sensitivity
-from .sensitivity_n import sensitivity_n
+from .sensitivity_n import sensitivity_n, seg_sensitivity_n
 from .deletion_until_flip import deletion_until_flip
-from .seg.irof import irof
-from .seg.seg_sensitivity_n import seg_sensitivity_n
-from .seg.iiof import iiof
+from .irof_iiof import irof, iiof
